create external router for API Gateway
