+++ localstack/http/router.py
@@ -126,7 +126,9 @@ class Router(Generic[E]):
     def __init__(
         self, dispatcher: Dispatcher[E] = None, converters: Mapping[str, Type[BaseConverter]] = None
     ):
-        self.url_map = Map(host_matching=True, strict_slashes=False, converters=converters)
+        self.url_map = Map(
+            host_matching=True, strict_slashes=False, converters=converters, redirect_defaults=False
+        )
         self.dispatcher = dispatcher or call_endpoint
         self._mutex = threading.RLock()
